+++ packages/adminui-pub/locale/lang-az.js
@@ -129,6 +129,9 @@ UB.i18nExtend({
   singOperationInProcess: 'Başqa sənədin imzalanmasını Running. Sonra hərəkətə təkrar',
 
   // all messages below are formatted with 4 params [0=UB.i18n(featureInfo.UIName), 1=featureInfo.minVersion, 2=currentVersion, 3=installerDownloadPath]
+  NMInstallExtensionFirefox: '<p>İşi davam etdirmək üçün Firefox brauzeri üçün <b>"UBExtension"</b> genişlənməsini quraşdırmaq lazımdır.</p> ' +
+    '<p> "Google" mağazasına keçmək üçün <a href="https://addons.mozilla.org/addon/ub-extension/" target="_blank">GENİŞLƏNMƏNİ QURAŞDIR</a> keçidi üzərində vurun</p>' +
+    '<p>Genişlənməni quraşdırdıqdan sonra <b>brauzeri söndürüb yandırın!</b></p>',
   NMInstallExtensionChrome: '<p>İşi davam etdirmək üçün "Google Chrome" brauzeri üçün <b>"UBExtension"</b> genişlənməsini quraşdırmaq lazımdır.</p> ' +
         '<p> "Google" mağazasına keçmək üçün <a href="https://chrome.google.com/webstore/detail/ubextension/{3}" target="_blank">GENİŞLƏNMƏNİ QURAŞDIR</a> keçidi üzərində vurun</p>' +
         '<p> Mağaza pəncərəsində <img src="models/adminui-pub/ub-extension/ChromePlusFreeEn.png"/> düyməsini(ola bilər ki, <b>+ÖDƏNİŞSİZ</b>) sıxın </p>' +